Changed checkstyle on travis to the check goal.
Multiple fixes have been made to now actually let the build fail when
there is a checkstyle error. Checkstyle is invoked a different way and
it is invoked with arguments for the location of the custom_style file
and the suppression file, because it didn't find it before and used the
default.

Also the checkstylediff is adjusted to make it work better due to
presenting the file locations differently.

Lastly the supression markup is fixed to reflect the specifications of
the xml sheet used. Also the custom style file is adusted somewhat because
there were specification errors there before.