Added support for checkstyle suppression for legacy coders.
Code that does not adhere to our checkstyle rules is still
being developed while we develop code that must adhere to checkstyle
rules.
The situation came up where conflicts needed to be resolved of
new "legacy" code but because they are changes compared to the context
branch they were still deemed as needing proper checkstyling.

This commit creates the additional distinction, where legacy coders
work is still suppressed. This is done by finding the latest commit of any
legacy coder on the branch and checkstyle changes are only from that point
checked. The use case is that legacy coders first finish their work before
we work further on their changes, most of the time fixing conflicts
between their changes and the context branch.