Changed the way in which the latest legacy contributor is chosen.
Instead of taking the commit date, which does not necessarily dictate
that it is the latest commit of a legacy coder in the feature/develop
branch the following is done. The commit log diff from feature branch
compared to the context branch is taken and the latest commit is
chosen to through list order that is from a legacy contributor.