Changed custom actions implementation. (detailed log below)
Changed return value of CustomAction to a single Percept instead of a list.
Removed the returnsPercept function, null can also be returned anyways.
Added TygronEntity to CustomActions. Some actions might want to know the caller.
Renamed CounterTestAction to a less-work-in-progress name.
Made it possible for Actions to return a Percept to the EIS engine.
Small renames and refactors.